Use the same tox flake8 command on CI and locally
CI uses "flake8 ." while we have a tox -eflake8 job that returns a
different result and currently doesn't pass.

This patch removes the direct call to flake8 in the github action to
replace it with a call to the existing tox environment and fixes an
error currently returned by the tox job:

tests/unit/test_scaffold.py:41:31: E126 continuation line over-indented for hanging indent
tests/unit/test_scaffold.py:44:30: E126 continuation line over-indented for hanging indent

Additionally, make the flake8 CI job separate so that it is easier to
understand if a unit test or lint job failed.