Implement configuration flexibility.
Previously, this library supported a small, specific list of
configuration options.  This patch implements a more flexible, layered
approach.  We now take two levels of configuration, one for direct flask
configuration and adding thing to the app.config object as Flask would
typically do, and a second for custom configuration. This first version
supports custom configuration via cfg and yaml files, but in the future,
we can add support for others as needed (json, toml, etc).

Signed-off-by: Jason Guiditta <jguiditt@redhat.com>