Fix parsing of config values.
Previously, when parsing the config passed in by the user, we
were artificially throwing an error if the value of a given key was
not a string or dictionary.  This check was to decide if further
parsing was needed to open an additional file specified in the config.
This patch fixes that logic so an error is not thrown when a value that
is neither string nor dictionary is passed in.

Signed-off-by: Jason Guiditta <jguiditt@redhat.com>